requirements.txt
gradio>=4.0
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0